Adding more security

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
 	const {
 		method,
 		query: { id },
+		cookies,
 	} = req
+
+	const token = cookies.token
 
 	dbConnect()
 
@@ -19,6 +22,9 @@ export default async function handler(req, res) {
 	}
 
 	if (method === 'PUT') {
+		if (!token || token !== process.env.TOKEN) {
+			return res.statuc(401).json('Not Authenticated!');
+		}
 		try {
 			const product = await Product.findByIdAndUpdate(id, req.body, {
 				new: true
@@ -30,6 +36,9 @@ export default async function handler(req, res) {
 	}
 
 	if (method === 'DELETE') {
+		if (!token || token !== process.env.TOKEN) {
+			return res.statuc(401).json('Not Authenticated!');
+		}
 		try {
 			await Product.findByIdAndDelete(id)
 			res.status(201).json('The product has been deleted!')
